Added ./configure system for Linux and Mac Removed pre-packaged zlib and libjpeg packages -- use system versions instead Added start for libPlasmaNet (not yet finished)

+++ core/PRP/Region/plRelevanceRegion.cpp
@@ -1,6 +1,6 @@
 #include "plRelevanceRegion.h"
 
-plRelevanceRegion::plRelevanceRegion() : fMgrIdx(-1) { }
+plRelevanceRegion::plRelevanceRegion() : fMgrIdx((unsigned int)-1) { }
 plRelevanceRegion::~plRelevanceRegion() { }
 
 IMPLEMENT_CREATABLE(plRelevanceRegion, kRelevanceRegion, plObjInterface)
